+++ editxt/commands.py
@@ -46,7 +46,7 @@ from editxt.command.openfile import open_, open_files
 from editxt.command.pathfind import pathfind
 from editxt.command.python import python
 from editxt.command.sortlines import sort_lines
-from editxt.command.split import split_text
+from editxt.command.split import split_text, join_lines
 from editxt.command.unique import unique_lines
 from editxt.command.wraplines import wrap_at_margin, wrap_lines
 
@@ -70,6 +70,7 @@ def load_commands():
             wrap_at_margin,
             wrap_lines,
             split_text,
+            join_lines,
             sort_lines,
             unique_lines,
             reindent,